Material now working with link
All files must be in the same directory path in the migrated destination
as in the original. E.g. do the move in the original into a single
folder and migrate into a single folder on the desitnation which is at
the same file path. When opened in the diestination project, you may
then use the 'move' function to rearange files into approprate
locations.